Split CLI & Pipeline
